backend/requirements.txt
Flask
beautifulsoup4
requests
pinecone-client
openai
python-dotenv
validators